requirements.txt
requests
schema
jsane
pytest
pytest-allure-adaptor
pymongo
configparser
redis
arrow
tablib
